setup.py: Generate builtin_version.py in git only
The build system always removed builtin_version.py.
When distributing tarballs, though, we don't have .git
and thus can't generate this file during the build.

This changes things so that the builtin_version.py
module is only generated and deleted when inside
of a git repository.

+++ setup.py
@@ -25,10 +25,12 @@ def main():
         _check_pyqt_version()
         _build_translations()      # msgfmt: .po -> .qm
     try:
-        version.write_builtin_version()
+        if os.path.exists('.git'):
+            version.write_builtin_version()
         _run_setup(git_version)
     finally:
-        version.delete_builtin_version()
+        if os.path.exists('.git'):
+            version.delete_builtin_version()
     # restore the old mask
     os.umask(old_mask)
